requirements.txt
aniso8601==9.0.1
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Faker==20.0.3
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
